ISR's to 'change'
The encoder alternates magnetic poles, and the max9921 interprets them as 1-0-1-0-etc.  Both rising and falling edges are mechanically precise, and using both will reduce the output latency (since you get an extra pulse to measure compared to the output).

Filter changed to weight the previous value more heavily as a resut.