Improved error handling.
Prior commits did not utilize the error or diagnostics pins on the Max, and any error would blank the outputs long enough to throw an ABS error at the ECU.

New commit will automatically reset the Max9921 after reading an error. as quickly as possible (<2uS) Multiple fast errors (10s or less between errors) will disable the output and read out the error status on the VSS pin (if unused for vehicle speed). All errors are sent out on the serial bus.